Enhanced homepage with Swiper features
Refined and streamlined the inclusion of Google Fonts and Swiper library CSS for a cleaner look in the HTML head section. Introduced comments to clarify the purpose of external resources and scripts. Expanded Swiper functionality in the homepage slider by adding pagination, custom navigation buttons, and a custom scrollbar to enrich user interaction and provide a smoother navigation experience.