feat: logo to navigation bar
Included HomeVise's logo in the navigation bar at the top of the homepage for enhanced brand visibility and a quicker way for users to return to the homepage. The logo is positioned as the first item in the list of navigation links, ensuring it's the first element users see in the navigation bar. This change is part of ongoing efforts to improve site navigation and reinforce brand identity.