feat: Enhance theme with color variables and styles aligning with logo theme
Introduced CSS variables for theme colors to centralize control over the look and feel of the website, aligning it with the logo's color scheme. This change simplifies future adjustments and ensures consistency across the site. Removed the import for 'Open Sans' font to streamline font usage around 'Roboto'. Added background and text color variables to the body and html tags to apply the new color theme. Expanded styling to include header navigation with a complementary color scheme, and refined layout with flexbox for better alignment and spacing. Additionally, integrated styles for a responsive hamburger menu, though its functionality is hidden by default for future activation.