Added login and registration page
Introduced a new login and registration HTML page to enhance user accessibility to the site's features. This page includes separate sections for login and registration, complete with forms for email and password inputs, ensuring a straightforward path for both new and returning users to access their accounts or create new ones. Additionally, navigation improvements were made to facilitate easy movement between this and other key pages of the website.